git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport/XBMC@10096 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/DllLoader/dll_tracker_file.cpp
@@ -16,7 +16,7 @@
 #include <io.h>
 #endif
 
-extern "C" void tracker_file_track(unsigned long caller, unsigned handle, TrackedFileType type, const char* sFile)
+extern "C" void tracker_file_track(uintptr_t caller, unsigned handle, TrackedFileType type, const char* sFile)
 {
   DllTrackInfo* pInfo = tracker_get_dlltrackinfo(caller);
   if (pInfo)
@@ -29,7 +29,7 @@ extern "C" void tracker_file_track(unsigned long caller, unsigned handle, Tracke
   }
 }
 
-extern "C" void tracker_file_free(unsigned long caller, unsigned handle, TrackedFileType type)
+extern "C" void tracker_file_free(uintptr_t caller, unsigned handle, TrackedFileType type)
 {
   DllTrackInfo* pInfo = tracker_get_dlltrackinfo(caller);
   if (pInfo)
@@ -77,7 +77,7 @@ extern "C"
 {
   int track_open(const char* sFileName, int iMode)
   {
-    unsigned loc = (unsigned)_ReturnAddress();
+    uintptr_t loc = (uintptr_t)_ReturnAddress();
     
     int fd = dll_open(sFileName, iMode);
     if (fd >= 0) tracker_file_track(loc, fd, FILE_XBMC_OPEN, sFileName);
@@ -86,7 +86,7 @@ extern "C"
 
   int track_close(int fd)
   {
-    unsigned loc = (unsigned)_ReturnAddress();
+    uintptr_t loc = (uintptr_t)_ReturnAddress();
     
     tracker_file_free(loc, fd, FILE_XBMC_OPEN);
     return dll_close(fd);
@@ -94,7 +94,7 @@ extern "C"
 
   FILE* track_fopen(const char* sFileName, const char* mode)
   {
-    unsigned loc = (unsigned)_ReturnAddress();
+    uintptr_t loc = (uintptr_t)_ReturnAddress();
     
     FILE* fd = dll_fopen(_P(sFileName), mode);
     if (fd) tracker_file_track(loc, (unsigned)fd, FILE_XBMC_FOPEN, sFileName);
@@ -103,7 +103,7 @@ extern "C"
   
   int track_fclose(FILE* stream)
   {
-    unsigned loc = (unsigned)_ReturnAddress();
+    uintptr_t loc = (uintptr_t)_ReturnAddress();
     
     tracker_file_free(loc, (unsigned)stream, FILE_XBMC_FOPEN);
     return dll_fclose(stream);
@@ -111,7 +111,7 @@ extern "C"
   
   FILE* track_freopen(const char *path, const char *mode, FILE *stream)
   {
-    unsigned loc = (unsigned)_ReturnAddress();
+    uintptr_t loc = (uintptr_t)_ReturnAddress();
     
     tracker_file_free(loc, (unsigned)stream, FILE_XBMC_FOPEN);
     stream = dll_freopen(path, mode, stream);
